fixed bugs, optimized and tested

+++ aes.h
@@ -10,7 +10,7 @@ void aes_ecb_decrypt(uint8_t * outputb, uint8_t * inputb, size_t size, uint8_t *
 void aes_ctr_encrypt(uint8_t * outputb, uint8_t * inputb, size_t size, uint8_t * key, uint8_t * iv);
 #define aes_ctr_decrypt(outputb,inputb,size,key,iv) aes_ctr_encrypt(outputb,inputb,size,key,iv)
 
-void aes_ctr_encrypt_fr(uint8_t * outputb, uint8_t * inputb, size_t size, uint8_t * key, uint8_t * iv);
+void aes_ctr_encrypt_fr(uint8_t * outputb, uint8_t * inputb, int size, uint8_t * key, uint8_t * iv);
 #define aes_ctr_decrypt_fr(outputb,inputb,size,key,iv) aes_ctr_encrypt(outputb,inputb,size,key,iv)
 
 word_t fr_get_mask();
